fix: 1st review fixes
