+++ src/components/profile-form/profile-form.js
@@ -3,37 +3,35 @@ import React from "react";
 import {PasswordInput, EmailInput, Input, Button} from "@ya.praktikum/react-developer-burger-ui-components";
 import {useDispatch, useSelector} from "react-redux";
 import {checkToken, getUser, updateUser} from "../../services/actions";
+import {useForm} from "../../hooks/useForm";
 
 
 const ProfileForm = () => {
+
   const dispatch = useDispatch();
   const user = useSelector(store => store.getUser.user);
 
   React.useEffect(() => {
       dispatch(getUser());
       dispatch(checkToken());
-      setEmail(user.email);
-      setName(user.name);
+      inputs.setValues({name: user.name, email: user.email, password: ''});
     },
     [dispatch, user.email, user.name]);
 
-  const [email, setEmail] = React.useState(`${user.email}`);
-  const [name, setName] = React.useState(`${user.name}`);
-  const [password, setPassword] = React.useState('');
+  const inputs = useForm(``)
   const [change, setChange] = React.useState(false);
   const [disabled, setDisabled] = React.useState(true);
 
   const onSubmit = (e) => {
     e.preventDefault();
-    dispatch(updateUser(name, email));
+    dispatch(updateUser(inputs.values.name, inputs.values.email));
     setChange(false);
     setDisabled(true);
   }
 
   const onDiscard = (e) => {
     e.preventDefault();
-    setEmail(user.email);
-    setName(user.name);
+    inputs.setValues({name: user.name, email: user.email, password: ''});
     dispatch(getUser());
     setChange(false);
     setDisabled(true);
@@ -42,19 +40,21 @@ const ProfileForm = () => {
   return (
     <section>
       <form onSubmit={onSubmit} className={ProfileFormStyles.form}>
-        <Input value={name} type="text" onIconClick={() => setDisabled(false)} icon="EditIcon" name={'name'}
+        <Input value={inputs.values.name} type="text" onIconClick={() => setDisabled(false)} icon="EditIcon"
+               name={'name'}
                placeholder="Имя" onChange={(e) => {
-          setName(e.target.value);
+          inputs.handleChange(e);
           setChange(true)
         }} disabled={disabled}/>
-        <EmailInput value={email} name={'email'} placeholder="Логин" isIcon={true} onChange={(e) => {
-          setEmail(e.target.value);
+        <EmailInput value={inputs.values.email} name={'email'} placeholder="Логин" isIcon={true} onChange={(e) => {
+          inputs.handleChange(e);
           setChange(true)
         }}/>
-        <PasswordInput value={password} icon="EditIcon" name={'password'} placeholder="Пароль" onChange={(e) => {
-          setPassword(e.target.value);
-          setChange(true)
-        }}/>
+        <PasswordInput value={inputs.values.password} icon="EditIcon" name={'password'} placeholder="Пароль"
+                       onChange={(e) => {
+                         inputs.handleChange(e);
+                         setChange(true)
+                       }}/>
         {change && (<div className={ProfileFormStyles.buttons}>
           <Button htmlType="button" type="secondary" size="medium" onClick={onDiscard}>Отмена</Button>
           <Button htmlType="submit" type="primary" size="medium">Сохранить</Button>
