--- src/components/protected-route-element-auth/protected-route-element-auth.js
@@ -1,10 +1,0 @@
-import {Navigate} from 'react-router-dom';
-import {homeUrl} from "../../utils/constants";
-import {useSelector} from "react-redux";
-
-
-export function ProtectedRouteElementAuth({element}) {
-  const auth = useSelector(store => store.auth.isAuth);
-
-  return auth ? <Navigate to={homeUrl} state={{homePage: true}} replace/> : element;
-}
