+++ src/pages/error-404/error-404.js
@@ -4,12 +4,10 @@ import Error404Styles from "./error-404.module.css";
 
 const Error404Page = () => {
   return (
-    <>
-      <div className={Error404Styles.container}>
-        <h1 className={Error404Styles.title}>ОШИБКА 404</h1>
-        <p>В таком безграничном космическом пространстве мы не смогли найти то, что Вам нужно...</p>
-      </div>
-    </>
+    <div className={Error404Styles.container}>
+      <h1 className={Error404Styles.title}>ОШИБКА 404</h1>
+      <p>В таком безграничном космическом пространстве мы не смогли найти то, что Вам нужно...</p>
+    </div>
   );
 }
