+++ src/pages/login/login.js
@@ -5,9 +5,7 @@ import LoginForm from "../../components/login-form/login-form";
 const LoginPage = () => {
 
   return (
-    <>
-      <LoginForm/>
-    </>
+    <LoginForm/>
   );
 }
