+++ src/pages/main/main.js
@@ -2,45 +2,20 @@ import React from "react";
 import AppStyles from "./main.module.css";
 import BurgerIngredients from "../../components/burger-ingredients/burger-ingredients";
 import BurgerConstructor from "../../components/burger-constructor/burger-constructor";
-import Modal from "../../components/modal/modal";
-import IngredientDetails from "../../components/ingredient-details/ingredient-details";
 import {DndProvider} from "react-dnd";
 import {HTML5Backend} from "react-dnd-html5-backend";
 import {useDispatch, useSelector} from "react-redux";
-import {getIngredients, closeIngredientInfo, checkToken} from "../../services/actions";
-import {useLocation, useNavigate} from "react-router-dom";
-import IngredientDetailsPage from "../ingredient-details/ingredient-details";
-import {homeUrl} from "../../utils/constants";
+import {checkToken} from "../../services/actions";
 
 
 const MainPage = () => {
 
   const data = useSelector(store => store.ingredients);
   const dispatch = useDispatch();
-  const location = useLocation();
-  const navigate = useNavigate();
 
   React.useEffect(() => {
     dispatch(checkToken());
   }, [dispatch]);
-
-  React.useEffect(
-    () => {
-      dispatch(getIngredients());
-    },
-    [dispatch]
-  );
-
-  const handleCloseIngredient = () => {
-    dispatch(closeIngredientInfo());
-    navigate(`${homeUrl}`, {
-      state: {
-        homePage: true
-      }
-    });
-  };
-
-  const ingredient = JSON.parse(localStorage.getItem('ingredient'));
 
   return (
     <DndProvider backend={HTML5Backend}>
@@ -50,30 +25,11 @@ const MainPage = () => {
       {!data.ingredientsRequest &&
         !data.ingredientsFailed &&
         data.ingredients.length &&
-        <>
-          <main className={AppStyles.main}>
-
-            {!location.state?.homePage && !location.state?.modalOpen && ingredient && <IngredientDetailsPage/>}
-
-            {location.state?.modalOpen && (
-              <>
-                <Modal onClose={() => handleCloseIngredient()} header={'Детали ингредиента'}>
-                  <IngredientDetails/>
-                </Modal>
-
-                <BurgerIngredients/>
-                <BurgerConstructor/>
-              </>
-            )}
-
-            {(location.state?.homePage || !ingredient) && (
-              <>
-                <BurgerIngredients/>
-                <BurgerConstructor/>
-              </>
-            )}
-          </main>
-        </>}
+        <main className={AppStyles.main}>
+          <BurgerIngredients/>
+          <BurgerConstructor/>
+        </main>
+      }
     </DndProvider>
   );
 }
