+++ src/services/actions/ws-actions.js
@@ -1,0 +1,31 @@
+export const WS_CONNECTION_START = 'WS_CONNECTION_START';
+export const WS_CONNECTION_SUCCESS = 'WS_CONNECTION_SUCCESS';
+export const WS_CONNECTION_ERROR = 'WS_CONNECTION_ERROR';
+export const WS_CONNECTION_CLOSED = 'WS_CONNECTION_CLOSED';
+export const WS_GET_DATA = 'WS_GET_DATA';
+
+
+export const wsConnectionSuccess = () => {
+  return {
+    type: WS_CONNECTION_SUCCESS
+  };
+};
+
+export const wsConnectionError = () => {
+  return {
+    type: WS_CONNECTION_ERROR
+  };
+};
+
+export const wsConnectionClosed = () => {
+  return {
+    type: WS_CONNECTION_CLOSED
+  };
+};
+
+export const wsGetData = message => {
+  return {
+    type: WS_GET_DATA,
+    payload: message
+  };
+};
